Rename header file to match class name
