+++ Framework/Kernel/inc/MantidKernel/SplittingInterval.h
